hack: Add autopinning feature.
This commit adds a feature which I'm calling 'autopin'. This feature
is difficult to describe, which is why it's going to be a hack for the
foreseeable future. I'm also not really sure this is a sound feature.

Basically, manual pins for launchables (y) is subtracted from the user defined
number for autopins given by the user (x) and those pins will be automatically
added based on usage.